@crazycs520 2025-04-07
> will `send by chunk` make performance worse if the data-set is small?

I'm not sure what small data-set you mean, 10000 rows? If it's too small, I don't feel like it's making much difference.

IMO, `import into` is usually used in cases where the amount of data is relatively large.